fast_forward + read_output

+++ ocr/ocr.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void feed_forward(network *n, double *inputs);
-double read_output(network *n);
+size_t read_output(network *n);
